feat: create useNews hook
