+++ src/components/Navbar.jsx
@@ -4,7 +4,7 @@ export const Navbar = ({setCategory}) => {
     return (
         <nav className="navbar navbar-expand-lg bg-body-tertiary">
             <div className="container">
-                <a className="navbar-brand">Z1 News</a>
+                <a className="navbar-brand">T1 News</a>
                 <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                     <span className="navbar-toggler-icon"></span>
                 </button>
